@@ -9,18 +9,20 @@
     <packaging>war</packaging>
     <name>${project.artifactId}</name>
     <description>This is the basic Java EE application used in the demo.
-    It is a simple CRUD application. It uses Maven and Java EE 8 (JAX-RS, EJB, CDI, JSON-B, JSF, Bean Validation).
+    It is a simple CRUD application. It uses Maven and Jakarta EE 10 (JAX-RS, EJB, CDI, JSON-B, JSF, Bean Validation).
     </description>
     <properties>
-        <jakartaee.api.version>8.0.0</jakartaee.api.version>
-        <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-        <maven.compiler.target>1.8</maven.compiler.target>
-        <maven.compiler.source>1.8</maven.compiler.source>
-        <failOnMissingWebXml>false</failOnMissingWebXml>
+        <admin.group.id></admin.group.id>
+        <admin.group.name></admin.group.name>
         <client.id></client.id>
         <client.secret></client.secret>
+        <failOnMissingWebXml>false</failOnMissingWebXml>
+        <jakartaee.api.version>10.0.0</jakartaee.api.version>
+        <keystore.pass></keystore.pass>
+        <maven-war-plugin.version>3.4.0</maven-war-plugin.version>
+        <maven.compiler.release>17</maven.compiler.release>
+        <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <tenant.id></tenant.id>
-        <admin.group.id></admin.group.id>
     </properties>
     <dependencies>
         <dependency>
@@ -30,9 +32,9 @@
             <scope>provided</scope>
         </dependency>
         <dependency>
-            <groupId>javax.xml.bind</groupId>
-            <artifactId>jaxb-api</artifactId>
-            <version>2.3.1</version>
+            <groupId>jakarta.xml.bind</groupId>
+            <artifactId>jakarta.xml.bind-api</artifactId>
+            <version>4.0.0</version>
             <scope>provided</scope>
         </dependency>
         <dependency>
@@ -63,6 +65,41 @@
     </dependencies>
     <build>
         <finalName>${project.artifactId}</finalName>
+        <plugins>
+            <plugin>
+                <groupId>org.apache.maven.plugins</groupId>
+                <artifactId>maven-compiler-plugin</artifactId>
+                <version>3.8.1</version>
+                <configuration>
+                    <release>17</release>
+                </configuration>
+            </plugin>
+            <plugin>
+                <groupId>org.apache.maven.plugins</groupId>
+                <artifactId>maven-war-plugin</artifactId>
+                <version>${maven-war-plugin.version}</version>
+            </plugin>            
+            <plugin>
+                <groupId>org.apache.maven.plugins</groupId>
+                <artifactId>maven-enforcer-plugin</artifactId>
+                <version>3.0.0-M3</version>
+                <executions>
+                    <execution>
+                        <id>enforce-java</id>
+                        <goals>
+                            <goal>enforce</goal>
+                        </goals>
+                        <configuration>
+                            <rules>
+                                <requireJavaVersion>
+                                    <version>17</version>
+                                </requireJavaVersion>
+                            </rules>
+                        </configuration>
+                    </execution>
+                </executions>
+            </plugin>
+        </plugins>
     </build>
     <profiles>
         <profile>
@@ -76,7 +113,7 @@
                     <plugin>
                         <groupId>io.openliberty.tools</groupId>
                         <artifactId>liberty-maven-plugin</artifactId>
-                        <version>3.2</version>
+                        <version>3.9</version>
                         <executions>
                             <execution>
                                 <id>package-server</id>
