@@ -1,29 +1,29 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
     <artifactId>javaee-cafe-mysql</artifactId>
     <groupId>javaee-cafe</groupId>
-    <version>1.0.0</version>
+    <version>1.0.0-SNAPSHOT</version>
     <packaging>war</packaging>
     <name>${project.artifactId}</name>
     <description>This is the basic Java EE application used in the demo.
-    It is a simple CRUD application. It uses Maven and Java EE 8 (JAX-RS, EJB, CDI, JSON-B, JSF, Bean Validation, JPA).
+    It is a simple CRUD application. It uses Maven and Jakarta EE 10 (JAX-RS, EJB, CDI, JSON-B, JSF, Bean Validation).
     </description>
     <properties>
         <failOnMissingWebXml>false</failOnMissingWebXml>
+        <jakartaee.api.version>10.0.0</jakartaee.api.version>
+        <maven-war-plugin.version>3.4.0</maven-war-plugin.version>
+        <maven.compiler.release>17</maven.compiler.release>
+        <param.db.name>${env.DB_NAME}</param.db.name>
+        <param.db.password>${env.DB_PASSWORD}</param.db.password>
+        <param.db.port.number>${env.DB_PORT_NUMBER}</param.db.port.number>
+        <param.db.server.name>${env.DB_SERVER_NAME}</param.db.server.name>
+        <param.db.user>${env.DB_USER}</param.db.user>
+        <param.namespace>${env.NAMESPACE}</param.namespace>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
-        <maven.compiler.target>1.8</maven.compiler.target>
-        <maven.compiler.source>1.8</maven.compiler.source>
-        <jakartaee.api.version>8.0.0</jakartaee.api.version>
-        <maven-war-plugin.version>3.4.0</maven-war-plugin.version>
-        <param.db.server.name>${env.DB_SERVER_NAME}</param.db.server.name>
-        <param.db.port.number>${env.DB_PORT_NUMBER}</param.db.port.number>
-        <param.db.name>${env.DB_NAME}</param.db.name>
-        <param.db.user>${env.DB_USER}</param.db.user>
-        <param.db.password>${env.DB_PASSWORD}</param.db.password>
-        <param.namespace>${env.NAMESPACE}</param.namespace>
     </properties>
     <dependencies>
         <dependency>
@@ -65,6 +65,34 @@
             </resource>
         </resources>
         <plugins>
+            <plugin>
+                <groupId>org.apache.maven.plugins</groupId>
+                <artifactId>maven-compiler-plugin</artifactId>
+                <version>3.8.1</version>
+                <configuration>
+                    <release>17</release>
+                </configuration>
+            </plugin>
+            <plugin>
+                <groupId>org.apache.maven.plugins</groupId>
+                <artifactId>maven-enforcer-plugin</artifactId>
+                <version>3.0.0-M3</version>
+                <executions>
+                    <execution>
+                        <id>enforce-java</id>
+                        <goals>
+                            <goal>enforce</goal>
+                        </goals>
+                        <configuration>
+                            <rules>
+                                <requireJavaVersion>
+                                    <version>17</version>
+                                </requireJavaVersion>
+                            </rules>
+                        </configuration>
+                    </execution>
+                </executions>
+            </plugin>        
             <plugin>
                 <artifactId>maven-resources-plugin</artifactId>
                 <version>3.2.0</version>
