@@ -5,37 +5,36 @@
     <modelVersion>4.0.0</modelVersion>
     <artifactId>javaee-cafe-postgres</artifactId>
     <groupId>javaee-cafe</groupId>
-    <version>1.0.0</version>
+    <version>1.0.0-SNAPSHOT</version>
     <packaging>war</packaging>
     <name>${project.artifactId}</name>
     <description>This is the basic Java EE application used in the demo.
-    It is a simple CRUD application. It uses Maven and Java EE 8 (JAX-RS, EJB, CDI, JSON-B, JSF, Bean Validation, JPA).
+    It is a simple CRUD application. It uses Maven and Jakarta EE 10 (JAX-RS, EJB, CDI, JSON-B, JSF, Bean Validation).
     </description>
     <properties>
-        <jakartaee.api.version>8.0.0</jakartaee.api.version>
+        <db.name></db.name>
+        <db.password></db.password>
+        <db.port.number></db.port.number>
+        <db.server.name></db.server.name>
+        <db.user></db.user>
+        <failOnMissingWebXml>false</failOnMissingWebXml>
+        <jakartaee.api.version>10.0.0</jakartaee.api.version>
+        <liberty.bootstrap.artifactId>${project.artifactId}</liberty.bootstrap.artifactId>
+        <liberty.bootstrap.db.name>${db.name}</liberty.bootstrap.db.name>
+        <liberty.bootstrap.db.password>${db.password}</liberty.bootstrap.db.password>
+        <liberty.bootstrap.db.port.number>${db.port.number}</liberty.bootstrap.db.port.number>
+        <liberty.bootstrap.db.server.name>${db.server.name}</liberty.bootstrap.db.server.name>
+        <liberty.bootstrap.db.user>${db.user}</liberty.bootstrap.db.user>
+        <maven-war-plugin.version>3.4.0</maven-war-plugin.version>
+        <maven.compiler.release>17</maven.compiler.release>
+        <param.db.name>${env.DB_NAME}</param.db.name>
+        <param.db.password>${env.DB_PASSWORD}</param.db.password>
+        <param.db.server.name>${env.DB_SERVER_NAME}</param.db.server.name>
+        <param.db.user>${env.DB_USER}</param.db.user>
+        <param.namespace>${env.NAMESPACE}</param.namespace>
+        <param.replicas>3</param.replicas>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
-        <maven.compiler.target>1.8</maven.compiler.target>
-        <maven.compiler.source>1.8</maven.compiler.source>
-        <failOnMissingWebXml>false</failOnMissingWebXml>
-        <db.server.name></db.server.name>
-        <db.port.number></db.port.number>
-        <db.name></db.name>
-        <db.user></db.user>
-        <db.password></db.password>
-        <liberty.bootstrap.artifactId>${project.artifactId}</liberty.bootstrap.artifactId>
-        <liberty.bootstrap.db.server.name>${db.server.name}</liberty.bootstrap.db.server.name>
-        <liberty.bootstrap.db.port.number>${db.port.number}</liberty.bootstrap.db.port.number>
-        <liberty.bootstrap.db.name>${db.name}</liberty.bootstrap.db.name>
-        <liberty.bootstrap.db.user>${db.user}</liberty.bootstrap.db.user>
-        <liberty.bootstrap.db.password>${db.password}</liberty.bootstrap.db.password>
-        <param.replicas>3</param.replicas>
-        <param.db.server.name>${env.DB_SERVER_NAME}</param.db.server.name>
-        <param.db.port.number>${env.DB_PORT_NUMBER}</param.db.port.number>
-        <param.db.name>${env.DB_NAME}</param.db.name>
-        <param.db.user>${env.DB_USER}</param.db.user>
-        <param.db.password>${env.DB_PASSWORD}</param.db.password>
-        <param.namespace>${env.NAMESPACE}</param.namespace>
     </properties>
     <dependencies>
         <dependency>
@@ -74,6 +73,34 @@
                     </resource>
                 </resources>
                 <plugins>
+                    <plugin>
+                        <groupId>org.apache.maven.plugins</groupId>
+                        <artifactId>maven-compiler-plugin</artifactId>
+                        <version>3.8.1</version>
+                        <configuration>
+                            <release>17</release>
+                        </configuration>
+                    </plugin>
+                    <plugin>
+                        <groupId>org.apache.maven.plugins</groupId>
+                        <artifactId>maven-enforcer-plugin</artifactId>
+                        <version>3.0.0-M3</version>
+                        <executions>
+                            <execution>
+                                <id>enforce-java</id>
+                                <goals>
+                                    <goal>enforce</goal>
+                                </goals>
+                                <configuration>
+                                    <rules>
+                                        <requireJavaVersion>
+                                            <version>17</version>
+                                        </requireJavaVersion>
+                                    </rules>
+                                </configuration>
+                            </execution>
+                        </executions>
+                    </plugin>        
                     <plugin>
                         <artifactId>maven-resources-plugin</artifactId>
                         <version>3.2.0</version>
