@@ -1,11 +1,12 @@
 package cafe.web.view;
 
 import java.io.Serializable;
-import javax.enterprise.context.SessionScoped;
-import javax.inject.Inject;
-import javax.ws.rs.client.ClientRequestContext;
-import javax.ws.rs.client.ClientRequestFilter;
-import javax.ws.rs.core.HttpHeaders;
+
+import jakarta.inject.Inject;
+import jakarta.enterprise.context.SessionScoped;
+import jakarta.ws.rs.client.ClientRequestContext;
+import jakarta.ws.rs.client.ClientRequestFilter;
+import jakarta.ws.rs.core.HttpHeaders;
 
 @SessionScoped
 public class CafeRequestFilter implements Serializable, ClientRequestFilter {
